refactor: update backward compatibillity for newer Fabric
