+++ src/main/java/github/devhrytsan/radialhotbar/RadialHotBarMod.java
@@ -19,9 +19,8 @@ import org.slf4j.LoggerFactory;
 
 //? fabric {
 import github.devhrytsan.radialhotbar.platform.fabric.FabricPlatform;
-
+import github.devhrytsan.radialhotbar.compatibility.FabricKeyHelperBridge;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
-import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 
 //?} neoforge {
 /*import github.devhrytsan.radialhotbar.platform.neoforge.NeoforgePlatform;
@@ -85,7 +84,7 @@ public class RadialHotBarMod {
 				KEYBIND_CATEGORY
 		);
 
-		KeyBindingHelper.registerKeyBinding(OPEN_RADIAL_MENU_KEY);
+		FabricKeyHelperBridge.register(OPEN_RADIAL_MENU_KEY);
 
 		ClientTickEvents.END_CLIENT_TICK.register(client -> {
 			RadialMenuController.INSTANCE.HandleUpdate(client);
